% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoW.check.R
\name{geoW.check}
\alias{geoW.check}
\title{geoW.check}
\usage{
geoW.check(geoW, variable, percentage = 30, SIM, OUTLIER)
}
\arguments{
\item{geoW}{}

\item{variable}{}

\item{percentage}{}

\item{SIM}{}

\item{OUTLIER}{}
}
\description{
geoW.check
}
\examples{
geoW.check()
}
\keyword{geoW}

