% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoW.ncdf42geoW.R
\name{geoW.ncdf42geoW}
\alias{geoW.ncdf42geoW}
\title{geoW.ncdf42geoW}
\usage{
geoW.ncdf42geoW(path_ncdf, path_results, ncdf_type, filename_1, variable_file,
  variable_ncdf, UTM, adm_border, timestep, time_resolution,
  origin = "2014-03-05")
}
\arguments{
\item{path_ncdf}{}

\item{ncdf_type}{}

\item{filename_1}{}

\item{variable_file}{}

\item{variable_ncdf}{}

\item{UTM}{}

\item{adm_border}{}

\item{timestep}{}

\item{time_resolution}{}

\item{origin}{"2014-03-05", tz = "GMT", ok for eraInterim netcdf4 from Norwegian Meteorological Institute convert from ECMWF grib}
}
\description{
geoW.ncdf42geoW
}
\examples{
geoW.ncdf42geoW()
}
\keyword{geoW}

