% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoW.raster2xyp_R.R
\name{geoW.raster2xyp_R}
\alias{geoW.raster2xyp_R}
\title{geoW.raster2xyp_R}
\usage{
geoW.raster2xyp_R(b)
}
\arguments{
\item{b}{}
}
\description{
geoW.raster2xyp_R
}
\examples{
geoW.raster2xyp_R()
}
\keyword{geoW}

