% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoW.read_ncdf4.R
\name{geoW.read_ncdf4}
\alias{geoW.read_ncdf4}
\title{geoW.read_ncdf4}
\usage{
geoW.read_ncdf4(path, file_name, variable, west_coord, east_coord, south_coord,
  north_coord, origin = "2014-03-05")
}
\arguments{
\item{path}{}

\item{file_name}{}

\item{variable}{}

\item{west_coord}{}

\item{east_coord}{}

\item{south_coord}{}

\item{north_coord}{}

\item{origin}{"2014-03-05", tz = "GMT", ok for eraInterim netcdf4 from Norwegian Meteorological Institute convert from ECMWF grib}
}
\description{
geoW.read_ncdf4
}
\examples{
geoW.read_ncdf4()
}
\keyword{geoW}

